feat: some services

+++ src/controllers/tripController.ts
@@ -1,9 +1,14 @@
 import { Request, Response } from "express";
 import { TripObject } from "../protocols/protocols.js";
 import { deleteTripRepository, getAllTripRepository, getTripRepository, postTripRepository } from "../repositories/tripRepository.js";
+import { testDate, tripExists } from "../services/tripService.js";
 
 export async function postTrip(req: Request, res: Response){
     const trip = req.body as TripObject;
+    if(testDate(trip.date)){
+        res.status(400).send({message: "The date must be in the future"});
+        return;
+    }
 
     try{
         await postTripRepository(trip);
@@ -16,6 +21,10 @@ export async function postTrip(req: Request, res: Response){
 
 export async function deleteTrip(req: Request, res: Response){
     const id = Number(req.params.id) as number;
+    if(await tripExists(id)){
+        res.status(404).send({message: "Trip not found"});
+        return;
+    }
 
     try{
         await deleteTripRepository(id);
@@ -28,6 +37,10 @@ export async function deleteTrip(req: Request, res: Response){
 
 export async function getTrip(req: Request, res: Response){
     const id = Number(req.params.id) as number;
+    if(await tripExists(id)){
+        res.status(404).send({message: "Trip not found"});
+        return;
+    }
 
     try{
         const trip = await getTripRepository(id);
